# apis/openapi.yaml
openapi: 3.1.0
info:
  title: VY-PAM License Server API
  version: 1.0.0
  description: |
    Issuing, revocation, validation and platform settings for VY-PAM
    (see `VY-PAM_Enterprise_PAM_Architecture.md`).

    Served by `backend/phase2_license_server`. Conventions:

    - All business endpoints live under `/api/v1`; `/health` is the
      unauthenticated liveness probe.
    - Errors are JSON: `{"error": "<message>", "details": {...}}`.
    - Admin operations require the admin token as
      `Authorization: Bearer <token>` **or** `X-Admin-Token: <token>`.
      When `LICENSE_ADMIN_TOKEN` is unset the server runs in open/dev mode
      and every route is allowed (responses then carry `X-Auth-Mode: open`).
    - Writes that name an actor accept an `X-Actor` header (settings).
    - List endpoints paginate with `limit` / `offset` (limit max 200).
    - Timestamps are ISO 8601 strings.
servers:
  - url: http://127.0.0.1:5000
    description: Local development server (LICENSE_SERVER_HOST/PORT)
...
paths:
...
  /api/v1/licenses:
    get:
      tags: [licenses]
      summary: List licenses
      parameters:
        - name: status
          in: query
          schema: {type: string, enum: [active, revoked, expired]}
        - name: license_type
          in: query
          schema: {type: string}
        - name: issued_to
          in: query
          schema: {type: string}
        - name: tier
          in: query
          schema: {type: string}
        - name: q
          in: query
          description: Free-text search across key, ID and subject.
          schema: {type: string}
        - {$ref: "#/components/parameters/Limit"}
        - {$ref: "#/components/parameters/Offset"}
      responses:
        "200":
          description: Page of licenses
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LicenseList"}
        "400": {$ref: "#/components/responses/BadRequest"}
  /api/v1/licenses/import:
    post:
      tags: [licenses]
      summary: Import a vendor-signed license
      description: >
        Installs the signed .lic from the delivery bundle. The server verifies
        the signature against its trusted public key and stores the claims
        exactly as signed — it never signs (issuing belongs to PAM-MASTER).
        Expired, structurally invalid or already-installed files are refused.
      security:
        - bearerAuth: []
        - adminTokenHeader: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LicenseEnvelope"}
      responses:
        "201":
          description: License verified and stored
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ImportLicenseResponse"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
  /api/v1/licenses/{license_key}:
    get:
      tags: [licenses]
      summary: License detail
      description: Full record including usage summary and audit events.
      parameters:
        - {$ref: "#/components/parameters/LicenseKey"}
      responses:
        "200":
          description: One license
          content:
            application/json:
              schema:
                type: object
                required: [license]
                properties:
                  license: {$ref: "#/components/schemas/LicenseRecord"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/licenses/{license_key}/file:
    get:
      tags: [licenses]
      summary: Download the signed license file
      security:
        - bearerAuth: []
        - adminTokenHeader: []
        - {}
      parameters:
        - {$ref: "#/components/parameters/LicenseKey"}
        - name: format
          in: query
          description: Defaults to the format the license was signed with.
          schema: {type: string, enum: [json, jwt]}
      responses:
        "200":
          description: Signed envelope (json) or compact token (jwt) as an attachment
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LicenseEnvelope"}
            text/plain:
              schema: {type: string, description: Compact `header.payload.signature` token}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/licenses/{license_key}/revoke:
    post:
      tags: [licenses]
      summary: Revoke a license
      security:
        - bearerAuth: []
        - adminTokenHeader: []
        - {}
      parameters:
        - {$ref: "#/components/parameters/LicenseKey"}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: {type: string}
      responses:
        "200":
          description: License revoked
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LicenseMutationResponse"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/licenses/{license_key}/restore:
    post:
      tags: [licenses]
      summary: Restore a revoked license
      security:
        - bearerAuth: []
        - adminTokenHeader: []
        - {}
      parameters:
        - {$ref: "#/components/parameters/LicenseKey"}
      responses:
        "200":
          description: License restored to active
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LicenseMutationResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/licenses/{license_key}/usage:
    post:
      tags: [licenses]
      summary: Report runtime usage
      description: |
        A deployed instance reports consumption. Only the known usage fields
        plus `pools` (node-quota pools by id) are accepted; values are
        non-negative integers. The report is appended to the audit trail.
      security:
        - bearerAuth: []
        - adminTokenHeader: []
        - {}
      parameters:
        - {$ref: "#/components/parameters/LicenseKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UsageReportRequest"}
      responses:
        "201":
          description: Usage recorded
          content:
            application/json:
              schema:
                type: object
                required: [license, usage, message]
                properties:
                  license: {$ref: "#/components/schemas/LicenseRecord"}
                  usage: {$ref: "#/components/schemas/UsageSummary"}
                  message: {type: string}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/licenses/validate:
    post:
      tags: [validation]
      summary: Validate a license payload
      description: |
        Full check: structure -> cryptographic signature -> revocation list ->
        expiry. Accepts a JSON envelope, a `{license: ...}` / `{token: ...}`
        wrapper, or a compact token as a raw string body
        (`curl --data-binary @license.lic`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Envelope, wrapper, or the compact token as a JSON string.
              anyOf:
                - {$ref: "#/components/schemas/LicenseEnvelope"}
                - type: object
                  properties:
                    license: {type: object, additionalProperties: true}
                  additionalProperties: true
                - type: object
                  properties:
                    token: {type: string}
                  additionalProperties: true
                - type: string
          text/plain:
            schema: {type: string, description: Compact token (three dot-separated segments)}
      responses:
        "200":
          description: Verdict (also returned for invalid licenses)
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ValidationResult"}
        "400": {$ref: "#/components/responses/BadRequest"}
  /api/v1/licenses/{license_key}/check:
    post:
      tags: [validation]
      summary: Check feature / module / quota access
      description: Ask any combination of `feature`, `module_id` (one of the 8 catalog ids) and `limit_type` + `current_usage` in one call.
      parameters:
        - {$ref: "#/components/parameters/LicenseKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                feature: {type: string}
                module_id:
                  type: string
                  description: One of the 8 entitlement module ids.
                limit_type: {type: string}
                current_usage: {type: integer, minimum: 0}
              additionalProperties: false
      responses:
        "200":
          description: Per-dimension verdict
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AccessCheck"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /api/v1/settings/audit:
    get:
      tags: [settings]
      summary: Configuration changelog
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 0, maximum: 200, default: 20}
      responses:
        "200":
          description: Settings events, newest first
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SettingsAuditResponse"}
        "400": {$ref: "#/components/responses/BadRequest"}
  /api/v1/settings/{group}:
    put:
      tags: [settings]
      summary: Update one settings group
      description: |
        Merges the given fields into the group after strict validation
        (unknown fields/groups, wrong types, out-of-range values, malformed
        enums/URLs/bucket names are rejected with 400). A no-op write returns
        `{"message": "No changes"}` and records no event. Every accepted
        change appends a `SettingsEvent` with the actor (`X-Actor` header,
        else the authenticated admin) and a per-field old/new diff.
      security:
        - bearerAuth: []
        - adminTokenHeader: []
        - {}
      parameters:
        - name: group
          in: path
          required: true
          schema: {type: string, enum: [sso, hsm, zsp, worm]}
        - name: X-Actor
          in: header
          description: Who is making the change (recorded in the audit changelog).
          schema: {type: string, maxLength: 64}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              description: One or more fields of the group (see `schema` in GET /api/v1/settings).
              additionalProperties: true
      responses:
        "200":
          description: Stored (or no changes)
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SettingsWriteResponse"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /api/v1/events:
    get:
      tags: [ops]
      summary: Unified recent activity feed
      description: Recent entries across the license, settings, vault and discovery audit trails, newest first.
      parameters:
        - name: source
          in: query
          schema: {type: string, enum: [license, settings, vault, discovery]}
        - name: limit
          in: query
          schema: {type: integer, minimum: 0, maximum: 200, default: 20}
      responses:
        "200":
          description: Events newest first across the selected trails
          content:
            application/json:
              schema: {$ref: "#/components/schemas/EventFeed"}
        "400": {$ref: "#/components/responses/BadRequest"}
...
  /api/v1/vault/items:
    get:
      tags: [vault]
      summary: List inventory rows
      parameters:
        - name: q
          in: query
          description: Free-text search across name, description, target and principal.
          schema: {type: string}
        - name: type
          in: query
          schema:
            type: string
            enum: [database, cloud_iam, ssh_key, service_account, domain_password, api_token]
        - name: status
          in: query
          schema:
            type: string
            enum: [available, checked_out, rotating, rotation_due, failed]
        - {$ref: "#/components/parameters/Limit"}
        - {$ref: "#/components/parameters/Offset"}
      responses:
        "200":
          description: Page of inventory rows
          content:
            application/json:
              schema: {$ref: "#/components/schemas/VaultItemList"}
        "400": {$ref: "#/components/responses/BadRequest"}
    post:
      tags: [vault]
      summary: Onboard a credential
      description: Validates the payload strictly, stamps the first rotation window and records an audit event.
      security:
        - bearerAuth: []
        - adminTokenHeader: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VaultOnboardRequest"}
      responses:
        "201":
          description: Credential added to the inventory
          content:
            application/json:
              schema: {$ref: "#/components/schemas/VaultActionResponse"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /api/v1/vault/items/{item_id}:
    get:
      tags: [vault]
      summary: Inventory row detail with its audit trail
      parameters:
        - name: item_id
          in: path
          required: true
          schema: {type: integer, minimum: 1}
      responses:
        "200":
          description: One credential plus its recent events
          content:
            application/json:
              schema: {$ref: "#/components/schemas/VaultItemDetail"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/vault/items/{item_id}/checkout:
    post:
      tags: [vault]
      summary: Check a credential out (JIT)
      security:
        - bearerAuth: []
        - adminTokenHeader: []
        - {}
      parameters:
        - name: item_id
          in: path
          required: true
          schema: {type: integer, minimum: 1}
        - name: X-Actor
          in: header
          description: Who is checking the credential out (recorded in the audit trail).
          schema: {type: string, maxLength: 64}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: {type: string, maxLength: 255}
      responses:
        "200":
          description: Credential checked out
          content:
            application/json:
              schema: {$ref: "#/components/schemas/VaultActionResponse"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/vault/items/{item_id}/revoke:
    post:
      tags: [vault]
      summary: End an active checkout
      security:
        - bearerAuth: []
        - adminTokenHeader: []
        - {}
      parameters:
        - name: item_id
          in: path
          required: true
          schema: {type: integer, minimum: 1}
        - name: X-Actor
          in: header
          description: Who revokes the checkout (recorded in the audit trail).
          schema: {type: string, maxLength: 64}
      responses:
        "200":
          description: Checkout ended, credential available again
          content:
            application/json:
              schema: {$ref: "#/components/schemas/VaultActionResponse"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /api/v1/vault/events:
    get:
      tags: [vault]
      summary: Vault audit trail
      description: Onboarding, checkouts, revocations and rotations, newest first.
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 0, maximum: 200, default: 20}
        - name: action
          in: query
          description: Return only one action type (unknown values are rejected with 400).
          schema:
            type: string
            enum: [onboarded, checked_out, revoked, rotated, rotation_failed]
      responses:
        "200":
          description: Vault events
          content:
            application/json:
              schema: {$ref: "#/components/schemas/VaultEventFeed"}
        "400": {$ref: "#/components/responses/BadRequest"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Admin token as `Authorization: Bearer <token>`."
    adminTokenHeader:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: The same admin token in a custom header (either is accepted).
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 0, maximum: 200, default: 50}
    Offset:
      name: offset
      in: query
      schema: {type: integer, minimum: 0, default: 0}
...
  schemas:
...
    LicenseList:
      type: object
      required: [licenses, total, limit, offset]
      properties:
        licenses:
          type: array
          items: {$ref: "#/components/schemas/LicenseRecord"}
        total: {type: integer}
        limit: {type: integer}
        offset: {type: integer}
...